+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 [v3.1.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.1.0.0)
 - fix bug where the existing instance would not be brought to the front when an additional instance of werkbank was launched
+- add delete button
 
 [v3.0.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.0.0.0)
 - first release 🎉